framecraft/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "framecraft"
version = "0.0.1"
description = "Turn natural-language situations into Hyperframes projects."
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = { text = "MIT" }
authors = [{ name = "Kartik" }]
dependencies = [
    "pydantic>=2.6",
    "jinja2>=3.1",
    "typer[all]>=0.12",
    "rich>=13",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "pyyaml>=6",
    "platformdirs>=4",
    # LLM provider SDKs are hard deps; the adapter layer imports them lazily
    # so missing SDK errors only surface if that provider is actually selected.
    "google-genai>=0.3",
    "anthropic>=0.39",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-recording>=0.13",
    "ruff>=0.6",
    "mypy>=1.11",
]

[project.scripts]
framecraft = "framecraft.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/framecraft"]

[tool.hatch.build.targets.wheel.sources]
"src/framecraft" = "framecraft"

[tool.hatch.build.targets.wheel.force-include]
"src/framecraft/prompts/primer.md" = "framecraft/prompts/primer.md"
"src/framecraft/prompts/primer.lock.json" = "framecraft/prompts/primer.lock.json"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RET"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-ra --strict-markers"
markers = [
    "llm: tests that require a real LLM provider API key",
]
